initial dial respects delay/jitter
