+++ nats-base-client/types.ts
@@ -14,8 +14,7 @@
  */
 //@ts-ignore
 import { NatsConnection } from "./nats.ts";
-import { extend } from "./util.ts";
-import { ErrorCode, NatsError } from "./mod.ts";
+import { NatsError } from "./mod.ts";
 
 export const CLOSE_EVT = "close";
 
@@ -55,7 +54,7 @@ export interface ConnectionOptions {
   token?: string;
   url: string;
   user?: string;
-  userJWT?: () => string | string;
+  userJWT?: (() => string) | string;
   verbose?: boolean;
 
   maxPingOut?: number;
@@ -81,65 +80,6 @@ export interface TlsOptions {
   // these may not be supported on all environments
   caFile?: string;
   keyFile?: string;
-}
-
-export function defaultOptions(): ConnectionOptions {
-  return {
-    maxPingOut: DEFAULT_MAX_PING_OUT,
-    maxReconnectAttempts: DEFAULT_MAX_RECONNECT_ATTEMPTS,
-    noRandomize: false,
-    payload: Payload.STRING,
-    pedantic: false,
-    pingInterval: DEFAULT_PING_INTERVAL,
-    reconnect: true,
-    reconnectJitter: DEFAULT_JITTER,
-    reconnectJitterTLS: DEFAULT_JITTER_TLS,
-    reconnectTimeWait: DEFAULT_RECONNECT_TIME_WAIT,
-    tls: undefined,
-    verbose: false,
-    waitOnFirstConnect: false,
-  } as ConnectionOptions;
-}
-
-export function parseOptions(opts?: ConnectionOptions): ConnectionOptions {
-  opts = opts || { url: DEFAULT_URI };
-  if (opts.port) {
-    opts.url = DEFAULT_PRE + opts.port;
-  }
-
-  let options = extend(defaultOptions(), opts);
-  if (options.user && options.token) {
-    throw NatsError.errorForCode(ErrorCode.BAD_AUTHENTICATION);
-  }
-
-  let payloadTypes = [Payload.JSON, Payload.STRING, Payload.BINARY];
-  if (opts.payload && !payloadTypes.includes(opts.payload)) {
-    throw NatsError.errorForCode(ErrorCode.INVALID_PAYLOAD_TYPE);
-  }
-
-  ["nonceSigner", "reconnectDelayHandler"].forEach((n) => {
-    if (options[n] && typeof options[n] !== "function") {
-      throw new NatsError(
-        `${n} option should be a function`,
-        ErrorCode.NOT_FUNC,
-      );
-    }
-  });
-
-  if (!options.reconnectDelayHandler) {
-    options.reconnectDelayHandler = () => {
-      let extra = options.tls
-        ? options.reconnectJitterTLS
-        : options.reconnectJitter;
-      if (extra) {
-        extra++;
-        extra = Math.floor(Math.random() * extra);
-      }
-      return options.reconnectTimeWait + extra;
-    };
-  }
-
-  return options;
 }
 
 export interface Msg {
