+++ tests/helpers/mod.ts
@@ -1,4 +1,4 @@
-export { delay, check } from "./delay.ts";
+export { check } from "./delay.ts";
 export { Lock } from "./lock.ts";
 export { TestServer, Connection } from "./test_server.ts";
 export { assertErrorCode } from "./asserts.ts";
